+++ app/src/main/java/it/gym/config/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package it.goodfellas.config;
+package it.gym.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
