+++ app/src/main/java/it/gym/controller/SaleController.java
@@ -1,18 +1,18 @@
-package it.goodfellas.controller;
+package it.gym.controller;
 
-import it.goodfellas.exception.InvalidSaleException;
-import it.goodfellas.exception.SalesIsNotCompletedException;
-import it.goodfellas.exception.SalesLineItemNotDeletedException;
-import it.goodfellas.hateoas.SaleAssembler;
-import it.goodfellas.hateoas.SaleResource;
-import it.goodfellas.model.*;
-import it.goodfellas.repository.AdminRepository;
-import it.goodfellas.repository.SaleRepository;
-import it.goodfellas.repository.SalesLineItemRepository;
-import it.goodfellas.service.CustomerService;
-import it.goodfellas.service.SaleService;
-import it.goodfellas.service.SalesLineItemService;
-import it.goodfellas.service.TrainingBundleSpecificationService;
+import it.gym.exception.InvalidSaleException;
+import it.gym.exception.SalesIsNotCompletedException;
+import it.gym.exception.SalesLineItemNotDeletedException;
+import it.gym.hateoas.SaleAssembler;
+import it.gym.hateoas.SaleResource;
+import it.gym.model.*;
+import it.gym.repository.AdminRepository;
+import it.gym.repository.SaleRepository;
+import it.gym.repository.SalesLineItemRepository;
+import it.gym.service.CustomerService;
+import it.gym.service.SaleService;
+import it.gym.service.SalesLineItemService;
+import it.gym.service.TrainingBundleSpecificationService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
