+++ app/src/main/java/it/gym/controller/TimeOffController.java
@@ -1,17 +1,17 @@
-package it.goodfellas.controller;
+package it.gym.controller;
 
-import it.goodfellas.exception.InvalidTimesOff;
-import it.goodfellas.exception.TimesOffNotFound;
-import it.goodfellas.exception.UnAuthorizedException;
-import it.goodfellas.hateoas.TimeOffAssembler;
-import it.goodfellas.hateoas.TimeOffResource;
-import it.goodfellas.model.*;
-import it.goodfellas.repository.ReservationRepository;
-import it.goodfellas.repository.TimeOffRepository;
-import it.goodfellas.repository.TrainerRepository;
-import it.goodfellas.service.TimeOffService;
-import it.goodfellas.service.UserService;
-import it.goodfellas.utility.MailSenderUtility;
+import it.gym.exception.InvalidTimesOff;
+import it.gym.exception.TimesOffNotFound;
+import it.gym.exception.UnAuthorizedException;
+import it.gym.hateoas.TimeOffAssembler;
+import it.gym.hateoas.TimeOffResource;
+import it.gym.model.*;
+import it.gym.repository.ReservationRepository;
+import it.gym.repository.TimeOffRepository;
+import it.gym.repository.TrainerRepository;
+import it.gym.service.TimeOffService;
+import it.gym.service.UserService;
+import it.gym.utility.MailSenderUtility;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
