+++ app/src/main/java/it/gym/controller/TrainingReservationController.java
@@ -1,21 +1,21 @@
-package it.goodfellas.controller;
+package it.gym.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import it.goodfellas.exception.*;
-import it.goodfellas.hateoas.ReservationAssembler;
-import it.goodfellas.hateoas.ReservationResource;
-import it.goodfellas.hateoas.TrainingSessionAssembler;
-import it.goodfellas.hateoas.TrainingSessionResource;
-import it.goodfellas.model.*;
-import it.goodfellas.repository.TimeOffRepository;
-import it.goodfellas.repository.TrainerRepository;
-import it.goodfellas.repository.TrainingSessionRepository;
-import it.goodfellas.repository.UserRepository;
-import it.goodfellas.service.CustomerService;
-import it.goodfellas.service.ReservationService;
-import it.goodfellas.service.TrainingBundleService;
-import it.goodfellas.utility.MailSenderUtility;
+import it.gym.exception.*;
+import it.gym.hateoas.ReservationAssembler;
+import it.gym.hateoas.ReservationResource;
+import it.gym.hateoas.TrainingSessionAssembler;
+import it.gym.hateoas.TrainingSessionResource;
+import it.gym.model.*;
+import it.gym.repository.TimeOffRepository;
+import it.gym.repository.TrainerRepository;
+import it.gym.repository.TrainingSessionRepository;
+import it.gym.repository.UserRepository;
+import it.gym.service.CustomerService;
+import it.gym.service.ReservationService;
+import it.gym.service.TrainingBundleService;
+import it.gym.utility.MailSenderUtility;
 import org.apache.commons.lang3.time.DateUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,7 +38,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
-import javax.websocket.server.PathParam;
 import java.security.Principal;
 import java.util.*;
 import java.util.stream.Collectors;
