+++ app/src/main/java/it/gym/exception/AvailabilityCheckException.java
@@ -1,4 +1,4 @@
-package it.goodfellas.exception;
+package it.gym.exception;
 
 public class AvailabilityCheckException extends RuntimeException {
