+++ app/src/main/java/it/gym/exception/InvalidVerificationLink.java
@@ -1,4 +1,4 @@
-package it.goodfellas.exception;
+package it.gym.exception;
 
 public class InvalidVerificationLink extends RuntimeException {
     public InvalidVerificationLink(String verifyLink) {
