+++ app/src/main/java/it/gym/hateoas/AdminResource.java
@@ -1,6 +1,6 @@
-package it.goodfellas.hateoas;
+package it.gym.hateoas;
 
-import it.goodfellas.model.Admin;
+import it.gym.model.Admin;
 import org.springframework.hateoas.Resources;
 
 public class AdminResource extends ResourceModel<Admin> {
