+++ app/src/main/java/it/gym/hateoas/CustomerAssembler.java
@@ -1,8 +1,8 @@
-package it.goodfellas.hateoas;
+package it.gym.hateoas;
 
-import it.goodfellas.model.Customer;
-import it.goodfellas.repository.CustomerRepository;
-import it.goodfellas.utility.Constants;
+import it.gym.model.Customer;
+import it.gym.repository.CustomerRepository;
+import it.gym.utility.Constants;
 import org.springframework.hateoas.Resources;
 import org.springframework.hateoas.mvc.ResourceAssemblerSupport;
