+++ app/src/main/java/it/gym/hateoas/SaleAssembler.java
@@ -1,8 +1,8 @@
-package it.goodfellas.hateoas;
+package it.gym.hateoas;
 
-import it.goodfellas.controller.SaleController;
-import it.goodfellas.model.Sale;
-import it.goodfellas.utility.Constants;
+import it.gym.controller.SaleController;
+import it.gym.model.Sale;
+import it.gym.utility.Constants;
 import org.springframework.hateoas.Resources;
 import org.springframework.hateoas.mvc.ResourceAssemblerSupport;
